Refactoring of BIMTester Module (#1259)
