+++ src/ifcblenderexport/blenderbim/bim/__init__.py
@@ -17,6 +17,7 @@ if bpy is not None:
         "context": None,
         "covetool": None,
         "csv": None,
+        "bimtester": None,
         "debug": None,
         "geometry": None,
         "model": None,
@@ -37,7 +38,6 @@ if bpy is not None:
         operator.SelectClass,
         operator.SelectType,
         operator.OpenUri,
-        operator.SelectFeaturesDir,
         operator.SelectDiffJsonFile,
         operator.SelectDiffNewFile,
         operator.SelectDiffOldFile,
@@ -47,14 +47,8 @@ if bpy is not None:
         operator.ValidateIfcFile,
         operator.ExportIFC,
         operator.ImportIFC,
-        operator.ColourByClass,
         operator.ColourByAttribute,
         operator.ColourByPset,
-        operator.ResetObjectColours,
-        operator.ApproveClass,
-        operator.RejectClass,
-        operator.SelectAudited,
-        operator.RejectElement,
         operator.SelectExternalMaterialDir,
         operator.AddSweptSolid,
         operator.RemoveSweptSolid,
@@ -153,8 +147,6 @@ if bpy is not None:
         operator.PropagateTextData,
         operator.ConvertLocalToGlobal,
         operator.ConvertGlobalToLocal,
-        operator.ExecuteBIMTester,
-        operator.BIMTesterPurge,
         operator.SelectIfcPatchInput,
         operator.SelectIfcPatchOutput,
         operator.ExecuteIfcPatch,
@@ -252,7 +244,6 @@ if bpy is not None:
         ui.BIM_PT_constraints,
         ui.BIM_PT_search,
         ui.BIM_PT_ifcclash,
-        ui.BIM_PT_qa,
         ui.BIM_PT_library,
         ui.BIM_PT_gis,
         ui.BIM_PT_presentation_layers,
