+++ src/ifcblenderexport/blenderbim/bim/ui.py
@@ -1069,63 +1069,6 @@ class BIM_PT_search(Panel):
         row.operator("bim.colour_by_pset", text="", icon="BRUSH_DATA")
 
 
-class BIM_PT_qa(Panel):
-    bl_label = "BIMTester Quality Auditing"
-    bl_idname = "BIM_PT_qa"
-    bl_options = {"DEFAULT_CLOSED"}
-    bl_space_type = "PROPERTIES"
-    bl_region_type = "WINDOW"
-    bl_context = "scene"
-
-    def draw(self, context):
-        layout = self.layout
-        layout.use_property_split = True
-
-        scene = context.scene
-        bim_properties = bpy.context.scene.BIMProperties
-
-        layout.label(text="Gherkin Setup:")
-
-        row = layout.row(align=True)
-        row.prop(bim_properties, "features_dir")
-        row.operator("bim.select_features_dir", icon="FILE_FOLDER", text="")
-
-        if bim_properties.features_dir:
-            row = layout.row(align=True)
-            row.prop(bim_properties, "features_file")
-
-            row = layout.row(align=True)
-            row.prop(bim_properties, "scenario")
-        else:
-            return
-
-        row = layout.row()
-        row.operator("bim.execute_bim_tester")
-
-        row = layout.row()
-        row.operator("bim.bim_tester_purge")
-
-        layout.label(text="Quality Auditing:")
-
-        row = layout.row()
-        row.prop(bim_properties, "qa_reject_element_reason")
-        row = layout.row()
-        row.operator("bim.reject_element")
-
-        row = layout.row(align=True)
-        row.operator("bim.colour_by_class")
-        row.operator("bim.reset_object_colours")
-
-        row = layout.row()
-        row.prop(bim_properties, "audit_ifc_class")
-
-        row = layout.row(align=True)
-        row.operator("bim.approve_class")
-        row.operator("bim.reject_class")
-
-        row = layout.row()
-        row.operator("bim.select_audited")
-
 
 class BIM_PT_library(Panel):
     bl_label = "IFC BIM Server Library"
